% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cctcTBSS.R
\name{cctcTBSS}
\alias{cctcTBSS}
\title{This function implement TreeScan for a case-case-time control design}
\usage{
cctcTBSS(
  case,
  control,
  tree,
  B = 999,
  parallel = FALSE,
  ncpus = NULL,
  nodeToRemove = NULL,
  seed = 8277
)
}
\arguments{
\item{`case`:}{a matrix or data.frame with three columns (leavesID, cases in the reference window, cases in the hazard window)}

\item{`control`:}{a matrix or data.frame with three columns (leavesID, cases in the reference window, cases in the hazard window)}

\item{`tree`:}{a matrix or data.frame with two columns (nodeID, parantNodeID)}

\item{`B`}{number of bootstrap replicates. Default is `999`.}

\item{`parallel`}{if TRUE parallel computing is used to compute the bootstrap p-values.}

\item{`ncpus`}{number of cpus used if `parallel = TRUE`. If no value is provided set to  `detectCores() - 1`.}
}
\description{
This function implement TreeScan for a case-case-time control design
}
