% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{computeLRT}
\alias{computeLRT}
\title{Compute node-specific discrepancy measures}
\usage{
computeLRT(object, parallel = FALSE, ncpus = NULL)
}
\arguments{
\item{object}{an object of class TS}

\item{parallel}{if \code{parallel = TRUE} parallel computing is used to compute the Monte Carlo p-values.}

\item{ncpus}{number of cpus used if \code{parallel = TRUE}. If no value is provided, the default is \code{ncpus = detectCores() - 1}.}
}
\description{
generic function to compute node-specific discrepancy measures. The function populates the slot \code{nodeSS}.
}
