% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalPower.R
\name{evalPower}
\alias{evalPower}
\title{Evaluate power of TBSS for a specific alternative distribution}
\usage{
evalPower(
  object,
  alternativeParameter = 0,
  alternativeLeaves = NULL,
  alpha = 0.05,
  R = 1000,
  parallel = FALSE,
  ncpus = NULL
)
}
\arguments{
\item{object}{an object of class \code{TS}}

\item{alternativeParameter}{either a number or a vector of lenght(alternativeLeaves) specifying the parameters to be used for the alternative model. For Poisson scan statistics this is an increase in relative risk, while for Bernoulli scan statistics is the probability under the alternative hypothesis.}

\item{alternativeLeaves}{a vector with the name of the leaves that will be generated under the alternative model. See details}

\item{alpha}{is the type-I error used for the test. The default is 0.05.}

\item{R}{is the number of dataset simulated under the alternative and used to estimate the power.}

\item{parallel}{if \code{TRUE} parallel computing is used to compute the bootstrap p-values.}

\item{ncpus}{number of cpus used if \code{parallel = TRUE}. If no value is provided set to  \code{ncpus = detectCores() - 1}.}
}
\description{
Evaluate power of TBSS for a specific alternative distribution
}
