% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLeaf.R
\name{getLeaf}
\alias{getLeaf}
\title{Retrieve leaves linked to a node}
\usage{
getLeaf(node, tree, leafs)
}
\arguments{
\item{`node`}{a character vector with the id of a node}

\item{`tree`:}{a data.frame with two columns. The first column is the id of each node of a tree while the second is its direct ancestor. Root node as no ancestor that is represented with "".}

\item{`leafs`:}{a list of the terminal nodes, i.e. the leaves}
}
\description{
Internal function that list the leveas connected to a node
}
