% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalPoissonTS.R
\name{H0_gen,conditionalPoissonTS-method}
\alias{H0_gen,conditionalPoissonTS-method}
\title{generate a new tree under the global null H0}
\usage{
\S4method{H0_gen}{conditionalPoissonTS}(object)
}
\description{
generate a new tree under the global null H0
}
