% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconditionalBernoulliTS.R
\name{H0_gen,unconditionalBernoulliTS-method}
\alias{H0_gen,unconditionalBernoulliTS-method}
\title{generate a new tree under the global null H0}
\usage{
\S4method{H0_gen}{unconditionalBernoulliTS}(object)
}
\description{
generate a new tree under the global null H0
}
