% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconditionalBernoulliTSVariableRR.R
\name{H0_gen,unconditionalBernoulliTSVariableRR-method}
\alias{H0_gen,unconditionalBernoulliTSVariableRR-method}
\title{generate data under the null}
\usage{
\S4method{H0_gen}{unconditionalBernoulliTSVariableRR}(object)
}
\description{
generate data under the null
}
