% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconditionalPoissonTS.R
\name{H0_gen,unconditionalPoissonTS-method}
\alias{H0_gen,unconditionalPoissonTS-method}
\title{generate data under the null}
\usage{
\S4method{H0_gen}{unconditionalPoissonTS}(object)
}
\description{
generate data under the null
}
