% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{H0_gen}
\alias{H0_gen}
\title{Generate data under the null distribution}
\usage{
H0_gen(object)
}
\arguments{
\item{object}{an object of class TS}
}
\description{
generic function to generate data under the null distribution. Instances of this function are used in \code{monteCarlo}
}
