% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{H1_gen}
\alias{H1_gen}
\title{Generate data under an alternative distribution}
\usage{
H1_gen(object, alternativeParameter, alternativeLeaves)
}
\arguments{
\item{object}{an object of class TS}

\item{alternativeParameter}{a number or a vector of length \code{length(alternativeLeaves)} containing parameters to be used to change the distribution of the \code{alternativeLeaves}}

\item{alternativeLeaves}{a vector containing the IDs of the leaves to be modified from the null distribution.}
}
\description{
generic function to generate data under an alternative distribution. The generation starts from the null and modifies generation parameters for the leaves in \code{alternativeLeaves}. The argument \code{alternativeParameter} contains parameters used to change the distribution of the \code{alternativeLeaves}. The meaning of this parameter varies according to the specific TBSS. Instances of this function are used to compute power in \code{evalPower}.
}
