% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{listLeaves}
\alias{listLeaves}
\title{Populate the leaves slot}
\usage{
listLeaves(object)
}
\arguments{
\item{object}{an object of class TS}
}
\description{
Populate the slot leaves of TS with a vector of the unique values of the first column of data (leaves ID).
}
