% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{listNodes}
\alias{listNodes}
\title{Populate the slot nodes}
\usage{
listNodes(object)
}
\arguments{
\item{object}{an object of class TS}
}
\description{
Populate the slot node of TS with a vector containing the unique value of the first column of the slot \code{tree}. Empty \code{character} are not allowed
}
