% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{monteCarlo}
\alias{monteCarlo}
\title{Monte Carlo algorithm}
\usage{
monteCarlo(object, parallel = FALSE, ncpus = NULL)
}
\arguments{
\item{object}{an object of class TS}

\item{parallel}{if \code{parallel = TRUE} parallel computing is used to compute the Monte Carlo p-values.}

\item{ncpus}{number of cpus used if \code{parallel = TRUE}. If no value is provided, the default is \code{ncpus = detectCores() - 1}.}
}
\description{
Function that approximates the null distribution of the node-specific discrepancy measures. Results are stored in the slot \code{LRT_H0}
}
