% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_llrt.R
\name{poisson_llrt}
\alias{poisson_llrt}
\title{compute the log-likelihood ratio test for two Poisson with different offsets (person time)}
\usage{
poisson_llrt(observed, observed_pt, expected, expected_pt)
}
\description{
compute the log-likelihood ratio test for two Poisson with different offsets (person time)
}
