% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonTBSS.R
\name{poissonTBSS}
\alias{poissonTBSS}
\title{Poisson TBSS}
\usage{
poissonTBSS(
  data,
  tree,
  min_events = 2,
  B = 999,
  parallel = FALSE,
  ncpus = NULL,
  type = "unconditional",
  direction = "positive",
  nodeToRemove = NULL,
  seed = 8277
)
}
\arguments{
\item{data}{a \code{matrix} or \code{data.frame} with three columns (leavesID, observed, expected).}

\item{tree}{a \code{matrix} or \code{data.frame} with two columns (nodeID, parentID) that specifies the hierarchical structure in an edge list format.}

\item{min_events}{the minimum number of observed events for a node to be tested. This is an integer greater or equal to 2. Default \code{min_events = 2}. Note that by choosing higher \code{min_events}, the power can increase because of the decreased number of hypotheses that are tested. This argument should be fixed before running the analysis to preserve the validity of the p-values.}

\item{B}{number of Monte Carlo replicates. The default is \code{B = 999}.}

\item{parallel}{if \code{parallel = TRUE} parallel computing is used to compute the Monte Carlo p-values.}

\item{ncpus}{number of cpus used if \code{parallel = TRUE}. If no value is provided, the default is \code{ncpus = detectCores() - 1}.}

\item{type}{determines the type of Poisson TBSS to perform. Possible values are \code{type = "unconditional"} for unconditional Poisson TBSS, and \code{type = "conditional"} for Poisson TBSS conditional to the total number of events in the tree.  Default is \code{type = "unconditional"}.}

\item{direction}{can be either \code{'positive'}, \code{"negative"} or, \code{"all"} by default (\code{direction = 'positive'}) poissonTBSS only look at nodes where the number of expected cases is larger than the number of observed cases, or equivalently the excess of cases is positive. Setting  \code{direction = '\code{"all"}'} test all nodes, while \code{direction  = "negative"} only node with negative excess of cases.}

\item{nodeToRemove}{a character vector with user-specified nodes to be removed from the analysis.  Note that sufficient statistics for these nodes are computed, but they are excluded from the testing procedure. By removing nodes from the list of tests, power can increase because of the decreased number of hypotheses that are tested. This list should be specified before running the analysis.}

\item{seed}{a number indicating the seed used for the analysis, i.e., the input of \code{set.seed()}. Default \code{seed  = 8277}.}
}
\description{
This function implements unconditional and conditional Poisson TBSS analysis.
}
\examples{
## Unconditional Poisson TBSS examples
library(dplyr)
library(TBSS)

# load tree 
data("tree_example", package = 'TBSS')

# load data
data("poisson_TBSS_examples", package = 'TBSS')

dat = poisson_TBSS_example
head(dat)

head(dat)


mod_tbss = poissonTBSS(data = dat,
                        tree = tree_example)
# inspect potential signals
summary(mod_tbss) |> filter(pvalue <=0.05)

mod_tbss = poissonTBSS(data = dat,
                      tree = tree_example,
                      direction = 'all')
# inspect potential signals
summary(mod_tbss) |> filter(pvalue <=0.05)

# conditional analysis
mod_tbss_cond = poissonTBSS(data = dat,
                        tree = tree_example,
                        type = 'conditional'  )
# inspect potential signals
summary(mod_tbss_cond) |> filter(pvalue <=0.05)
}
