% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionalPoissonTS.R
\name{summary,conditionalPoissonTS-method}
\alias{summary,conditionalPoissonTS-method}
\title{generate a new tree under the global null H0}
\usage{
\S4method{summary}{conditionalPoissonTS}(object)
}
\description{
generate a new tree under the global null H0
}
