% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconditionalBernoulliTSVariableRR.R
\name{summary,unconditionalBernoulliTSVariableRR-method}
\alias{summary,unconditionalBernoulliTSVariableRR-method}
\title{generate a new tree under the global null H0}
\usage{
\S4method{summary}{unconditionalBernoulliTSVariableRR}(object)
}
\description{
generate a new tree under the global null H0
}
