% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unconditionalPoissonTS.R
\name{summary,unconditionalPoissonTS-method}
\alias{summary,unconditionalPoissonTS-method}
\title{generate a new tree under the global null H0}
\usage{
\S4method{summary}{unconditionalPoissonTS}(object)
}
\description{
generate a new tree under the global null H0
}
