% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_example.R
\docType{data}
\name{tree_example}
\alias{tree_example}
\title{An example of tree with 25 nodes}
\format{
## `tree_example`
A data frame with 24 rows and 2 columns describing a hierarchical structure
\describe{
  \item{childID}{node ID from where an edge start.}
  \item{parentID}{node ID to where an edge end.}
}
}
\usage{
tree_example
}
\description{
The format of the data is the standard format used by TBSS with two columns, childID and parentID.
}
\keyword{datasets}
