% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\docType{class}
\name{TS}
\alias{TS}
\title{General TS class for TBSS}
\description{
General TS class for TBSS
}
\section{Slots}{

\describe{
\item{\code{data}}{contains the data in a \code{data.frame} one row for each leaf of the tree.  The number of columns varies depending on the analysis, while the first column is the name of a leaf.}

\item{\code{tree}}{a\code{data.frame} with two columns representing the hierarchical structure in a edge list format. Each row represents a directional edge between a node and its parent node, so the two columns are a nodeID and parentID. The IDs are treated as \code{character}.}

\item{\code{leaves}}{is a vector with the names of the leaves, typically the first column of the slot \code{data}}

\item{\code{nodes}}{a vector with the names of all the nodes in the tree, i.e. the first column of \code{tree}}

\item{\code{mapNodesLeaves}}{is a named list. Each element of the list is named after a node, while the content of the list indicates the leaves associated with that node.}

\item{\code{nodesToTest}}{is a vector defining the nodes to be tested. Some nodes could be excluded from the testing procedure  (e.g., because there are too few data points) and can be excluded from this list.}

\item{\code{nodeSS}}{is a \code{data.frame} reporting the sufficient statistics for each node in the tree. The number of columns varies depending on the analysis, and the first column is the name of a node.}

\item{\code{LRT}}{is a named vector containing the node-specific discrepancy measure computed on the observed data for each node in \code{nodesToTest}.}

\item{\code{B}}{is the number of MC replicates used to approximate the null distribution of LRT}

\item{\code{LRT_H0}}{is a \code{matrix} with the MC-distribution of LRT. Each column is an MC replicate.}
}}

