% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncon_bernoulli_llrt.R
\name{uncon_bernoulli_llrt}
\alias{uncon_bernoulli_llrt}
\title{compute the log-likelihood ratio test for unconditional binomial analysis}
\usage{
uncon_bernoulli_llrt(case, control, p)
}
\description{
compute the log-likelihood ratio test for unconditional binomial analysis
}
