% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncon_bernoulli_llrt.R
\name{uncon_bernoulli_llrt_vectorized}
\alias{uncon_bernoulli_llrt_vectorized}
\title{same function but vectorized.}
\usage{
uncon_bernoulli_llrt_vectorized(case, control, p)
}
\description{
same function but vectorized.
}
