% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncon_poisson_llrt.R
\name{uncon_poisson_llrt}
\alias{uncon_poisson_llrt}
\title{compute the log-likelihood ratio test for unconditional Poisson analysis}
\usage{
uncon_poisson_llrt(observed, expected)
}
\description{
compute the log-likelihood ratio test for unconditional Poisson analysis
}
